+++ boot/zephyr/CMakeLists.txt
@@ -113,6 +113,9 @@ zephyr_library_sources(
   ${BOOT_DIR}/bootutil/src/fault_injection_hardening.c
   )
 
+# library which might be common source code for MCUBoot and an application
+zephyr_link_libraries(MCUBOOT_BOOTUTIL)
+
 if(CONFIG_BOOT_FIH_PROFILE_HIGH)
 zephyr_library_sources(
   ${BOOT_DIR}/bootutil/src/fault_injection_hardening_delay_rng_mbedtls.c
