+++ boot/zephyr/include/mcuboot_config/mcuboot_logging.h
@@ -13,9 +13,13 @@
  * When building for targets running Zephyr, delegate to its native
  * logging subsystem.
  */
-
+#ifdef CONFIG_MCUBOOT
 #define MCUBOOT_LOG_MODULE_DECLARE(domain)	LOG_MODULE_DECLARE(domain, CONFIG_MCUBOOT_LOG_LEVEL)
 #define MCUBOOT_LOG_MODULE_REGISTER(domain)	LOG_MODULE_REGISTER(domain, CONFIG_MCUBOOT_LOG_LEVEL)
+#else
+#define MCUBOOT_LOG_MODULE_DECLARE(domain)	LOG_MODULE_DECLARE(domain, CONFIG_MCUBOOT_UTIL_LOG_LEVEL)
+#define MCUBOOT_LOG_MODULE_REGISTER(domain)	LOG_MODULE_REGISTER(domain, CONFIG_MCUBOOT_UTIL_LOG_LEVEL)
+#endif
 
 #define MCUBOOT_LOG_ERR(...) LOG_ERR(__VA_ARGS__)
 #define MCUBOOT_LOG_WRN(...) LOG_WRN(__VA_ARGS__)
